updates analysis currents RPCs CMS
These notebooks are used to analyze the current data of the RPCs in the CMS for the years 2018, 2022, and 2023. Within these notebooks, we introduce a new mapping function and an updated mapping dataframe that encompasses all the chambers in the CMS. These updates enable us to conduct a comprehensive analysis of the RPC currents.